Lukin i elem vec sum

+++ tasks/lukin_i_elem_vec_sum/common/include/common.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <cstddef>
+#include <cstdint>
 #include <string>
 #include <tuple>
 #include <vector>
@@ -10,8 +10,8 @@
 namespace lukin_i_elem_vec_sum {
 
 using InType = std::vector<int>;
-using OutType = int;
-using TestType = std::tuple<size_t, std::string>;
+using OutType = int64_t;
+using TestType = std::tuple<int, std::string>;
 using BaseTask = ppc::task::Task<InType, OutType>;
 
 }  // namespace lukin_i_elem_vec_sum
